+++ CHANGELOG.md
@@ -6,6 +6,8 @@
         - Allow cobrands to override searching by reference #2271
     - Front end improvements:
         - Clearer relocation options while you’re reporting a problem #2238
+    - Admin improvements
+        - List number of alerts on report page #669
     - Bugfixes:
         - Add perl 5.26/5.28 support.
         - Fix subcategory issues when visiting /report/new directly #2276